MNT: Move partial data multi-tau to accumulators
