+++ skxray/core/correlation/__init__.py
@@ -1,2 +1,1 @@
-from .correlation import (multi_tau_auto_corr, auto_corr_scat_factor,
-                          multi_tau_auto_corr_partial_data)
+from .correlation import (multi_tau_auto_corr, auto_corr_scat_factor)
